Move DocumentList to shared so it can be shared.
