+++ src/scenes/Office/DocumentViewer/index.jsx
@@ -12,7 +12,7 @@ import { loadMoveDependencies } from '../ducks.js';
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
 import PrivateRoute from 'shared/User/PrivateRoute';
 import { Switch, Redirect, Link } from 'react-router-dom';
-import DocumentList from './DocumentList';
+import DocumentList from 'shared/DocumentViewer/DocumentList';
 import DocumentUploader from './DocumentUploader';
 import DocumentDetailPanel from './DocumentDetailPanel';
 import DocumentUploadViewer from './DocumentUploadViewer';
